refactor: re organizing project
There is now 3 main modules: CLI, Discord and Cosmos; along with a main and errors at project root
