+++ src/cli/commands/start.rs
@@ -6,9 +6,7 @@ use abscissa_core::{config, Command, FrameworkError, FrameworkErrorKind, Runnabl
 use clap::Parser;
 use tracing::{error, info};
 
-use crate::config::{DiscordBotConfig, DiscordShardingSection};
-use crate::discord::client;
-use crate::prelude::*;
+use crate::{discord::client, cli::{prelude::*, config::{DiscordBotConfig, DiscordShardingSection}}};
 
 #[derive(Command, Debug, Parser)]
 #[clap(arg_required_else_help(true))]
