+++ src/cli/mod.rs
@@ -1,0 +1,7 @@
+//! Holds CLI libs
+
+pub mod application;
+pub mod prelude;
+mod config;
+mod commands;
+mod components;
