+++ src/cosmos/mod.rs
@@ -1,0 +1,3 @@
+//! Holds Cosmos related files
+
+pub mod grpc;
